backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pillow==10.2.0
django-filter==23.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
django-storages==1.14.2
boto3==1.34.34
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
gunicorn==21.2.0
whitenoise==6.6.0